// apps/desktop/src/generated/FileEntryDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FileEntryKind } from "./FileEntryKind";

export type FileEntryDto = {
  name: string;
  relativePath: string;
  kind: FileEntryKind;
  size: number | null;
  modifiedAt: number | null;
};